Changed the argument of the fetching to a command object
The fetching pipeline on a virtual host used a string for specifying
a certain fetch command. This didn't allow for (easy) addition of
parameters to the fetch request (like the id of the field to be
fetched). So the pipeline uses a fetch command object now which
contains the string as well as an object for parameters.